Complete fixed-width integer migration for COM3D2 and binary I/O
Move binary I/O positions, lengths, alignment values, and 7-bit encoded sizes to exact int64/int32 types. Keep host int only at required io, bufio, sort, and reflection API boundaries, and remove redundant same-width conversions.

Match COM3D2 C# System.Int32 wire fields with int32 for ANM property indexes, COL/PHY/PSK/TEX/timeline counts, MODEL morph indexes, and PRESET slot/RID data. Use int64 for ARC, NEI, ABA, ContentTable, type-tree, and serializer offsets, lengths, cursors, and recursion depth.

Validate every published editing-schema root against host-width integer fields and regenerate the affected COM3D2 ANM, MODEL, and PRESET schemas with signed 32-bit bounds.

Wire widths were checked against the COM3D2 DynamicBone.cs, DynamicSkirtBone.cs, ImportCM.cs, and MaidProp.cs sources and the KCES AMEventParameter.cs, AMBinaryMethod.cs, and DLLAnmParse.cs sources under game/.

Verified with go test -count=1 for serialization, COM3D2/KCES services, schemagen, and editing schemas; go vet for serialization and schemagen; and GOARCH=386 compile-only tests for the same target packages. Full go test ./... remains blocked by the pre-existing undefined jsonschema references in application/validation.go.

+++ serialization/COM3D2/arc/arc_hash_test.go
@@ -122,7 +122,7 @@ func TestHashConsistency(t *testing.T) {
 func TestManualHash(t *testing.T) {
 	h := NameHashUTF16("a")
 	if h != 0x89be207bddf13e3 {
-		t.Errorf("Hash mismatch for 'a': expected %x, got %x", 0x89be207bddf13e3, h)
+		t.Errorf("Hash mismatch for 'a': expected %x, got %x", uint64(0x89be207bddf13e3), h)
 	}
 	t.Logf("Hash of 'a': %x", h)
 }
